cago-server/requirements.txt
asgiref==3.5.2
astroid==2.12.12
attrs==22.1.0
black==22.10.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
coverage==6.5.0
dill==0.3.6
Django==4.1.2
django-cors-headers==3.13.0
django-extensions==3.2.1
django-filter==22.1
django-phonenumber-field==7.0.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.24.2
drf-standardized-errors==0.12.2
gunicorn==20.1.0
inflection==0.5.1
isort==5.10.1
jsonschema==4.17.0
lazy-object-proxy==1.7.1
mccabe==0.7.0
mypy-extensions==0.4.3
pathspec==0.10.1
phonenumberslite==8.13.0
platformdirs==2.5.2
psycopg2==2.9.4
PyJWT==2.6.0
pylint==2.15.5
pylint-django==2.5.3
pylint-plugin-utils==0.7
pyparsing==3.0.9
pyrsistent==0.19.2
python-decouple==3.6
pytz==2022.5
PyYAML==6.0
sqlparse==0.4.3
tomli==2.0.1
tomlkit==0.11.5
typing_extensions==4.4.0
uritemplate==4.1.1
wrapt==1.14.1
